# Conflicts:
#	src/main/java/at/qe/skeleton/api/services/MeasurementService.java

+++ src/main/java/at/qe/skeleton/ui/controllers/PlantController.java
@@ -5,6 +5,7 @@ import java.util.Collection;
 
 import at.qe.skeleton.model.Userx;
 import at.qe.skeleton.services.PlantService;
+import at.qe.skeleton.services.SensorStationService;
 import at.qe.skeleton.ui.beans.SessionInfoBean;
 import lombok.Getter;
 import lombok.Setter;
@@ -19,6 +20,9 @@ import org.springframework.stereotype.Component;
 public class PlantController implements Serializable{
     @Autowired
     private PlantService plantService;
+
+    @Autowired
+    private SensorStationService sensorStationService;
 
     @Autowired
     private SessionInfoBean sessionInfoBean;
@@ -37,10 +41,18 @@ public class PlantController implements Serializable{
         return plantService.getAllPlants();
     }
 
+
+    /**
+     * The method is Only used in the scrolldown menu for plant selection.
+     */
+    public Collection<String> getPlantsUniqueNames() {
+        return plantService.getAllPlantsUniqueNames();
+    }
+
     /**
      * Returns how many plants are registered in the system.
      */
-    public Integer getPlantsAmount() {
+    public Long getPlantsAmount() {
         return plantService.getPlantsAmount();
     }
 
@@ -70,9 +82,15 @@ public class PlantController implements Serializable{
     /**
      * Action to delete the currently cached Sensor Station.
      */
+    //TODO: implement stronger error handling
     public void doDeletePlant() {
-        this.plantService.deletePlant(plant);
-        plant = null;
+        if (plant.getSensorStation() == null) {
+            System.out.println("No sensor station assigned im deleting the plant");
+            plantService.detachAllImagesFromPlant(plant);
+            this.plantService.deletePlant(plant);
+            plant = null;
+        }
+        else System.out.println("cannot delete a plant that is currently displayed in a sensor station , im plant controller");
     }
